+++ services/discord_updates.py
@@ -1,0 +1,1 @@
+# Empty file for now,  but will be used to update discord channels with servers. But may not use